cli: remove pre_execute and post_execute and custom str func
`pre_execute` and `post_execute` serve little purpose, the same
functionality can be achieved by be calling whatever function
at the start or end of the `execute` function.
The custom `str` function causes problem when flags and comments are/
will be stored as references. The default `str` function provide the
same information